Break up TestHelpers
