+++ java/src/test/java/com/firebase/geofire/TestHelpers.java
@@ -1,0 +1,9 @@
+package com.firebase.geofire;
+
+public final class TestHelpers {
+    public static final long TIMEOUT_SECONDS = 5;
+
+    private TestHelpers() {
+        throw new AssertionError("No instances.");
+    }
+}
